+++ pdr_backend/lake/test/resources.py
@@ -4,11 +4,11 @@ from enforce_typing import enforce_types
 import polars as pl
 
 
-from pdr_backend.data_eng.constants import TOHLCV_COLS, TOHLCV_SCHEMA_PL
-from pdr_backend.data_eng.gql_data_factory import GQLDataFactory
-from pdr_backend.data_eng.model_data_factory import ModelDataFactory
-from pdr_backend.data_eng.ohlcv_data_factory import OhlcvDataFactory
-from pdr_backend.data_eng.plutil import (
+from pdr_backend.lake.constants import TOHLCV_COLS, TOHLCV_SCHEMA_PL
+from pdr_backend.lake.gql_data_factory import GQLDataFactory
+from pdr_backend.lake.model_data_factory import ModelDataFactory
+from pdr_backend.lake.ohlcv_data_factory import OhlcvDataFactory
+from pdr_backend.lake.plutil import (
     concat_next_df,
     initialize_df,
     transform_df,
